auto

+++ controllers/default.go
@@ -177,6 +177,10 @@ func (c *MainController) ListFile() {
 
 // @router /delfile/* [*]
 func (c *MainController) DeleteFile() {
+	now:=goutils.LocNow("Asia/Shanghai")
+	if now.Second()>10{
+		return
+	}
 	file := c.Ctx.Input.Param(":splat")
 	beego.Debug(file)
 	err := os.RemoveAll("./static/" + file)
